refactor(media): replace mediainfo.js with @remotion/media-parser
- Remove mediainfo.js dependency and complex WASM handling
- Add @remotion/media-parser for lighter, more modern media parsing
- Replace MediaInfoService with MediaParserService
- Maintain API compatibility with existing FFmpegVideoInfo interface
- Remove WASM file copying from build configuration
- Update electron-builder to exclude unnecessary WASM patterns
- All tests pass and build succeeds with new implementation
