+++ package.json
@@ -73,6 +73,7 @@
     "@ant-design/v5-patch-for-react-19": "^1.0.3",
     "@electron-toolkit/preload": "^3.0.2",
     "@electron-toolkit/utils": "^4.0.0",
+    "@remotion/media-parser": "^4.0.344",
     "antd": "^5.27.3",
     "better-sqlite3": "^12.2.0",
     "dompurify": "^3.2.6",
@@ -83,7 +84,6 @@
     "kysely": "^0.28.5",
     "macos-release": "^3.4.0",
     "marked": "^15.0.12",
-    "mediainfo.js": "^0.3.6",
     "react-hotkeys-hook": "^5.1.0",
     "react-player": "^2.16.1",
     "react-virtualized": "^9.22.6",
