+++ src/main/ipc.ts
@@ -24,7 +24,7 @@ import DictionaryService from './services/DictionaryService'
 import FFmpegService from './services/FFmpegService'
 import FileStorage from './services/FileStorage'
 import { loggerService } from './services/LoggerService'
-import MediaInfoService from './services/MediaInfoService'
+import MediaParserService from './services/MediaParserService'
 import NotificationService from './services/NotificationService'
 import { registerShortcuts, unregisterAllShortcuts } from './services/ShortcutService'
 import { themeService } from './services/ThemeService'
@@ -37,7 +37,7 @@ const logger = loggerService.withContext('IPC')
 const fileManager = new FileStorage()
 const dictionaryService = new DictionaryService()
 const ffmpegService = new FFmpegService()
-const mediaInfoService = new MediaInfoService()
+const mediaParserService = new MediaParserService()
 
 /**
  * Register all IPC handlers used by the main process.
@@ -444,15 +444,15 @@ export function registerIpc(mainWindow: BrowserWindow, app: Electron.App) {
     return ffmpegService.getFFmpegPath()
   })
 
-  // MediaInfo
+  // MediaParser (Remotion)
   ipcMain.handle(IpcChannel.MediaInfo_CheckExists, async () => {
-    return await mediaInfoService.checkMediaInfoExists()
+    return await mediaParserService.checkExists()
   })
   ipcMain.handle(IpcChannel.MediaInfo_GetVersion, async () => {
-    return await mediaInfoService.getMediaInfoVersion()
+    return await mediaParserService.getVersion()
   })
   ipcMain.handle(IpcChannel.MediaInfo_GetVideoInfo, async (_, inputPath: string) => {
-    return await mediaInfoService.getVideoInfo(inputPath)
+    return await mediaParserService.getVideoInfo(inputPath)
   })
 
   // shortcuts
